refactor: simplify options syntax for findAll'
